Task: ProfileComponent
